requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
flake8==5.0.4
idna==3.3
isort==5.10.1
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.20.0
requests==2.28.1
urllib3==1.26.12
vk-api==11.9.9
